[Refactor] TS로 변환 - 함수 매개변수에 타입 지정 - PageElements를 todoapp.d.ts로 이동
