+++ 02.typescript/todoapp/src/pages/fetch/fetchDetailData.ts
@@ -1,5 +1,7 @@
+import axios from "axios";
+
 /* fetchDetailData는 상세정보를 불러오는 함수 */
-const fetchDetailData = async (_id) => {
+const fetchDetailData = async (_id: number) => {
   try {
     const response = await axios(`http://localhost:33088/api/todolist/${_id}`);
     return response.data.item;
