+++ 02.typescript/todoapp/src/pages/info/TodoInfo.ts
@@ -2,20 +2,20 @@ import TodoUpdate from "../update/TodoUpdate";
 import fetchDetailData from "../fetch/fetchDetailData";
 import getErrorMessage from "../../utils/Error";
 
-const TodoInfo = async (_id: number) => {
+const TodoInfo = async ({_id} : {_id: string}) => {
   try {
     /* 1. 페이지 요소 생성 및 초기 세팅 */
     const pageElements = createPageElements();
 
     /* 2. 서버에 있는 TODO 상세 데이터 가져옴 */
-    const detailData = await fetchDetailData(_id);
+    const detailData = await fetchDetailData(Number(_id));
 
     /* 3. 서버에 있는 데이터 값을 각 요소에 넣어줌 */
     setData(pageElements, detailData);
 
     /* 4. Todo Update 수정, 저장 기능 */
     TodoUpdate(
-      _id,
+      Number(_id),
       pageElements.editButton,
       pageElements.deleteButton,
       pageElements.title,
@@ -40,12 +40,10 @@ const createPageElements = () => {
   const editButton = document.createElement("button");
 
   title.setAttribute("class", "title-input");
-  // 확인해보기(-)
   title.setAttribute("disabled", "true");
   title.setAttribute("maxlength", "25");
   createTime.setAttribute("class", "time");
   content.setAttribute("class", "content-textarea");
-  // 확인해보기(-)
   content.setAttribute("disabled", "true");
 
   detailContainer.setAttribute("class", "info-container");
@@ -73,15 +71,6 @@ const createPageElements = () => {
     deleteButton,
   };
 };
-
-interface PageElements {
-  detailContainer: HTMLDivElement;
-  title: HTMLInputElement;
-  createTime: HTMLDivElement;
-  content: HTMLTextAreaElement;
-  editButton: HTMLButtonElement;
-  deleteButton: HTMLButtonElement;
-}
 
 /* 받아온 데이터를 요소에 입력해주는 함수 */
 const setData = (pageElements: PageElements, detailData: TodoItem) => {
